♻ extract useFresco
